requirements.txt
# ─────────────────────────────────────────────
#  Taskverse – Python Dependencies
#  Compatible with Python 3.13 on Windows
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.0

# Telegram Bot
python-telegram-bot==21.9

# LLM
groq==0.13.0

# Database
motor==3.7.0
pymongo==4.12.1

# Scheduling
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
pydantic==2.11.3
pydantic-settings==2.8.1
httpx==0.27.0
pytz==2024.1
python-dateutil==2.9.0.post0
tenacity==9.1.2

# Logging
loguru==0.7.3

# Testing
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-mock==3.14.0